move core resolution logic into its own module
